Added real story to the game, and fixed small problems

+++ dojo/src/events.cairo
@@ -5,9 +5,9 @@ use starknet::ContractAddress;
 pub struct Decision {
     #[key]
     pub player: ContractAddress,
-    pub node_id: u16,
+    pub node_id: u32,
     pub choice: u8,
-    pub next_node: u16,
+    pub next_node: u32,
 }
 
 #[derive(Drop, Serde)]
@@ -15,14 +15,14 @@ pub struct Decision {
 pub struct StoryCompleted {
     #[key]
     pub player: ContractAddress,
-    pub final_node: u16,
+    pub final_node: u32,
 }
 
 #[derive(Drop, Serde, Copy)]
 #[dojo::event]
 pub struct InvalidChoice{
     #[key]
-    pub node_id: u16,
+    pub node_id: u32,
     pub text: felt252,
     pub choice: u8,
 }
